+++ lib/ethon/easies/options.rb
@@ -13,9 +13,9 @@ module Ethon
           :customrequest, :cainfo, :capath, :connecttimeout,
           :followlocation, :httpauth, :infilesize, :interface,
           :maxredirs, :nosignal, :postfieldsize, :copypostfields, :proxy,
-          :proxyauth, :proxytype, :timeout, :readdata, :sslcert, :ssl_verifypeer, :ssl_verifyhost,
-          :sslcerttype, :sslkey, :sslkeytype, :sslversion,
-          :url, :useragent, :userpwd, :verbose, :readfunction
+          :proxyauth, :proxyport, :proxytype, :timeout, :readdata, :sslcert,
+          :ssl_verifypeer, :ssl_verifyhost, :sslcerttype, :sslkey, :sslkeytype,
+          :sslversion, :url, :useragent, :userpwd, :verbose, :readfunction
         ])
         base.send(:attr_accessor, *Ethon::Easy::AVAILABLE_OPTIONS)
       end
@@ -64,7 +64,7 @@ module Ethon
         def int_options
           [
             :connecttimeout, :dns_cache_timeout, :infilesize, :maxredirs,
-            :postfieldsize, :ssl_verifyhost, :timeout
+            :postfieldsize, :proxyport, :ssl_verifyhost, :timeout
           ]
         end
       end